+++ Entity/RUTINA4.cs
@@ -11,6 +11,6 @@ namespace SIGED_API.Entity
 
         public int RUTINA_ID { get; set; }
 
-        public int CALIFICACION { get; set; }
+        public decimal CALIFICACION { get; set; }
     }
 }
